services/indexing-worker/pyproject.toml
[project]
name = "briefforge-indexing-worker"
version = "0.1.0"
description = "Background worker for BriefForge indexing pipeline (parse, chunk, embed, reindex)."
requires-python = ">=3.11"
dependencies = [
  "psycopg[binary,pool]==3.3.3",
  "redis==5.0.4",
  "rq==1.16.2",
  "pydantic==2.12.5",
  "python-dotenv==1.0.1",
  "openai==1.30.1",
  "tenacity==8.3.0",
]

[project.scripts]
indexing-worker = "src.main:main"

